podcast-transcript-scraper/requirements.txt
selenium==3.141.0
requests==2.25.0
boto3==1.16.35
airtable-python-wrapper==0.15.1
python-dotenv==0.17.0